+++ dependencies/os_layer/src/win32/K15_FileSystemWin32.cpp
@@ -37,11 +37,11 @@ char* K15_Win32ConvertToSystemPath(const char* p_FilePath)
 {
 	assert(p_FilePath);
 
-	uint32 filePathSize = strlen(p_FilePath);
+	size_t filePathSize = strlen(p_FilePath);
 	char* convertedFilePath = (char*)malloc(filePathSize + 1); //+1 for 0 terminator
 	const char* filePath = p_FilePath;
 
-	for (uint32 fileNameIndex = 0;
+	for (size_t fileNameIndex = 0;
 		 fileNameIndex < filePathSize;
 		 ++fileNameIndex)
 	{
